Fix error, move database file
